Added exposed_get_size

+++ DS/data_server.py
@@ -71,7 +71,8 @@ class DataService(rpyc.Service):
 
         # TODO: Handle exceptions
         def exposed_delete_file(self, file_path):
-            os.remove(file_path)
+            if os.path.isfile(DATA_DIR + file_path):
+                os.remove(file_path)
 
         def exposed_get_size(self, file_path: str):
             if os.path.isfile(DATA_DIR+file_path):
